+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.3
+
+* Rename `cacheFuture` to `loadCacheFuture` on `BoxerCacheHandler`
+* Rename `requestFuture` to `loadRequestFuture` on `BoxerCacheHandler`
+
 # 1.0.2
 
 * Enhance: Rename `isDisableCache` to `isEnableCache` on `BoxerCacheHandler`